1. Function getBestAvailableBitmap renamed 2. Some functions moved to DataUtils.kt
