+++ main_screen/src/main/res/values/dimens.xml
@@ -3,12 +3,12 @@
 
     <!-- MapsFragment begin -->
     <dimen name="interface_margin_top">10dp</dimen>
-    <dimen name="device_location_button_margin_vertical">20dp</dimen>
-    <dimen name="device_location_button_margin_end">10dp</dimen>
-    <dimen name="search_margin_bottom">20dp</dimen>
     <dimen name="search_margin_start">10dp</dimen>
+    <dimen name="search_margin_bottom">20dp</dimen>
     <dimen name="filter_screen_button_margin_horizontal">10dp</dimen>
     <dimen name="filter_screen_button_margin_bottom">20dp</dimen>
+    <dimen name="device_location_button_margin_vertical">20dp</dimen>
+    <dimen name="device_location_button_margin_end">10dp</dimen>
     <dimen name="buttons_icon_size">30dp</dimen>
     <!-- MapsFragment end -->
